Rename build folder from '/public' to '/docs' for integration with GH Pages
GitHub pages is somewhat limited in where you can store your source files.
They can either be on the gh-pages branch, or on the master branch in the following
folders:
/
/docs

Since the index.html folder in the root folder is just a template for webpack to build
the actual website, we must use the build folder as the root. Hence, we are changing
the build folder from /public to /docs.